requirements.txt
litgpt==0.4.6
accelerate==0.23.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
awscrt==0.20.11
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bitsandbytes==0.42.0
blessed==1.20.0
boto3==1.34.134
botocore==1.34.134
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.0
colorama==0.4.6
croniter==1.4.1
DataProperty==1.0.1
datasets==2.20.0
dateutils==0.6.12
decorator==5.1.1
deepdiff==6.3.1
dill==0.3.7
docopt==0.6.2
docstring-parser==0.15
editdistance==0.6.2
evaluate==0.4.0
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.100.0
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.12.2
fire==0.5.0
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
hf_transfer==0.1.6
httpcore==1.0.5
httpx==0.27.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
inquirer==3.1.3
ipadic==1.0.0
ipython==8.12.2
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jiwer==3.0.2
jmespath==1.0.1
joblib==1.3.2
jsonargparse==4.31.0
jsonlines==4.0.0
lightning==2.3.0
lightning-cloud==0.5.37
lightning-utilities==0.11.3.post0
lit==16.0.6
litdata==0.2.6
litserve==0.1.2
llvmlite==0.40.1
lm_eval==0.4.2
lxml==4.9.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mbstrdecoder==1.1.3
mdurl==0.1.2
mecab-python3==1.0.5
more-itertools==10.0.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.1
nltk==3.8.1
num2words==0.5.12
numba==0.57.1
numexpr==2.10.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
ordered-set==4.1.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathvalidate==3.2.0
peft==0.5.0
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.3.0
portalocker==2.8.2
prompt-toolkit==3.0.39
protobuf==4.25.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pybind11==2.13.1
pydantic==2.0.3
pydantic_core==2.3.0
Pygments==2.15.1
PyJWT==2.8.0
pytablewriter==1.2.0
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.6
pytorch-triton==0.0.1
pytz==2023.3
PyYAML==6.0.1
rapidfuzz==2.13.7
readchar==4.0.5
regex==2023.6.3
requests==2.32.3
responses==0.18.0
rich==13.4.2
rouge_score==0.1.2
s3transfer==0.10.2
sacrebleu==2.3.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.10.1
sentencepiece==0.2.0
sentence-transformers==2.4.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlitedict==2.1.0
stack-data==0.6.2
starlette==0.27.0
starsessions==1.3.0
sympy==1.12
tabledata==1.3.3
tabulate==0.9.0
tcolorpy==0.1.6
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tiktoken==0.4.0
timm==1.0.7
tokenizers==0.19.1
torch==2.3.0
torchmetrics==1.4.0.post0
torchvision==0.18.0
tqdm==4.66.4
tqdm-multiprocess==0.0.11
traitlets==5.9.0
transformers==4.42.1
triton==2.3.0
typeshed-client==2.3.0
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.30.1
wcwidth==0.2.6
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==3.0.3
word2number==1.1
xformers==0.0.29.post3
xxhash==3.2.0
yarl==1.9.2
zipp==3.16.2
zstandard==0.22.0
opencv-python
albumentations